requirements.txt
pyttsx3==2.98
pyaudio==0.2.14
requests==2.31.0
yfinance==0.2.36
pygame==2.5.2
python-dotenv==1.0.1
colorama==0.4.6
